+++ internal/grpcwrapper/rawtopic/rawtopicreader/rawtopicreader.go
@@ -3,6 +3,7 @@ package rawtopicreader
 import (
 	"errors"
 	"fmt"
+	"io"
 	"reflect"
 
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Topic"
@@ -30,6 +31,9 @@ func (s StreamReader) CloseSend() error {
 
 func (s StreamReader) Recv() (ServerMessage, error) {
 	grpcMess, err := s.Stream.Recv()
+	if xerrors.Is(err, io.EOF) {
+		return nil, err
+	}
 	if err != nil {
 		if !xerrors.IsErrorFromServer(err) {
 			err = xerrors.Transport(err)
@@ -46,7 +50,7 @@ func (s StreamReader) Recv() (ServerMessage, error) {
 		return nil, xerrors.WithStackTrace(fmt.Errorf("ydb: bad status from topic server: %v", meta.Status))
 	}
 
-	switch m := grpcMess.ServerMessage.(type) {
+	switch m := grpcMess.GetServerMessage().(type) {
 	case *Ydb_Topic.StreamReadMessage_FromServer_InitResponse:
 		resp := &InitResponse{}
 		resp.ServerMessageMetadata = meta
@@ -102,7 +106,7 @@ func (s StreamReader) Recv() (ServerMessage, error) {
 	default:
 		return nil, xerrors.WithStackTrace(fmt.Errorf(
 			"ydb: receive unexpected message (%v): %w",
-			reflect.TypeOf(grpcMess.ServerMessage),
+			reflect.TypeOf(grpcMess.GetServerMessage()),
 			ErrUnexpectedMessageType,
 		))
 	}