+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 * Fixed sometime panic on topic writer closing
 
+## v3.56.2
+* Fixed return private error for commit to stopped partition in topic reader.
+* Stopped wrapping err error as transport error at topic streams (internals)
+
 ## v3.56.1
 * Fixed fixenv usage (related to tests only)